#196 70%  beta published!

+++ src/com/taobao/rigel/rap/account/dao/AccountDao.java
@@ -194,6 +194,6 @@ public interface AccountDao {
 	 * @param teamId
 	 * @return
 	 */
-	List<Long> getUserIdList(int teamId);
+	List<Integer> getUserIdList(int teamId);
 
 }
